Commit until part4.21

+++ controllers/blogs.js
@@ -1,14 +1,32 @@
-const Blog = require('../models/blog')
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
+const Blog = require('../models/blog')
+const User = require('../models/user')
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.startsWith('Bearer ')) {
+    return authorization.replace('Bearer ', '')
+  }
+  return null
+}
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
 
   const body = request.body
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+ 
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
   let blog
 
   if(body.likes){
@@ -17,6 +35,7 @@ blogsRouter.post('/', async (request, response) => {
       author: body.author,
       url: body.url,
       likes: body.likes,
+      user: user._id
     })
 
   }else{
@@ -25,17 +44,43 @@ blogsRouter.post('/', async (request, response) => {
       author: body.author,
       url: body.url,
       likes: 0,
+      user: user._id
     })
   }
-
     const savedBlog = await blog.save()
+
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.status(201).json(savedBlog)
 
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+ 
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id).populate('user')
+
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' });
+  }
+
+  const blogUserID = blog.user._id.toString()
+
+  console.log(blogUserID)
+
+  if(decodedToken.id === blogUserID){
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }else{
+    return response.status(401).json({ error: 'unauthorized to delete' })
+  }
+  
 })
 
 blogsRouter.put('/:id', async (request, response) => {
